+++ example/lib/main.dart
@@ -4,7 +4,7 @@ import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_thermal_printer/flutter_thermal_printer.dart';
-import 'package:flutter_thermal_printer/utils/ble_device_window.dart';
+import 'package:flutter_thermal_printer/utils/printer.dart';
 
 void main() {
   runApp(const MyApp());
@@ -19,40 +19,33 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   final _flutterThermalPrinterPlugin = FlutterThermalPrinter.instance;
-  List<BleDeviceWindow> windowsBleList = [];
-  List<BluetoothDevice> bleDevices = [];
 
-  // Get Bluetooth devices list of Windows
-  Future<void> getWindowBleDevicesList() async {
-    try {
-      final devices =
-          await _flutterThermalPrinterPlugin.getWindowBleDevicesList();
-      setState(() {
-        windowsBleList = devices;
-      });
-    } on PlatformException {
-      print('Failed to get USB devices list.');
-    }
-  }
+  List<Printer> bleDevices = [];
 
-  // Get Bluetooth devices list of others
-  Future<void> getOthersBleDevicesList() async {
+  StreamSubscription<List<Printer>>? _devicesStreamSubscription;
+
+  //  Start scanning for BLE devices
+  Future<void> startScan() async {
     try {
       await _flutterThermalPrinterPlugin.startScan();
-      _flutterThermalPrinterPlugin.devicesStream.listen((event) {
+      _devicesStreamSubscription =
+          _flutterThermalPrinterPlugin.devicesStream.listen((event) {
         setState(() {
-          bleDevices = event.toSet().toList();
-          bleDevices.removeWhere((element) => element.platformName.isEmpty);
+          bleDevices = event.map((e) => Printer.fromJson(e.toJson())).toList();
+          bleDevices.removeWhere(
+            (element) => element.name == null || element.name!.isEmpty,
+          );
         });
       });
-    } on PlatformException {
-      print('Failed to get USB devices list.');
+    } catch (e) {
+      log('Failed to start scanning for devices $e');
     }
   }
 
   // Stop scanning for BLE devices
   Future<void> stopScan() async {
     try {
+      _devicesStreamSubscription?.cancel();
       await _flutterThermalPrinterPlugin.stopScan();
     } catch (e) {
       log('Failed to stop scanning for devices $e');
@@ -74,95 +67,46 @@ class _MyAppState extends State<MyApp> {
           children: [
             ElevatedButton(
               onPressed: () {
-                getWindowBleDevicesList();
-              },
-              child: const Text('Get Window Ble Devices List'),
-            ),
-            ElevatedButton(
-              onPressed: () {
                 if (bleDevices.isNotEmpty) {
                   stopScan();
                 }
-                getOthersBleDevicesList();
+                startScan();
               },
               child: const Text('Get Others Ble Devices List'),
             ),
             Expanded(
-              child: Row(
-                children: [
-                  Expanded(
-                    child: ListView.builder(
-                      itemCount: bleDevices.length,
-                      itemBuilder: (context, index) {
-                        return ListTile(
-                          onTap: () async {
-                            final isConnected =
-                                await _flutterThermalPrinterPlugin
-                                    .connect(bleDevices[index]);
-                            log("Devices: $isConnected");
-                          },
-                          title: Text(bleDevices[index].platformName),
-                          subtitle: Text(
-                              "VendorId: ${bleDevices[index].remoteId.str} - Connected: ${bleDevices[index].isConnected}"),
-                          trailing: IconButton(
-                            icon: const Icon(Icons.connect_without_contact),
-                            onPressed: () async {
-                              final profile = await CapabilityProfile.load();
-                              final generator =
-                                  Generator(PaperSize.mm80, profile);
-                              List<int> bytes = [];
-                              bytes += generator.text('Hello World');
-                              bytes += generator
-                                  .text("|||| FLUTTER THERMAL PRINTER ||||");
-                              bytes += generator.feed(2);
-                              bytes += generator.cut();
-                              await _flutterThermalPrinterPlugin.printData(
-                                bleDevices[index],
-                                bytes,
-                              );
-                            },
-                          ),
+              child: ListView.builder(
+                itemCount: bleDevices.length,
+                itemBuilder: (context, index) {
+                  return ListTile(
+                    onTap: () async {
+                      final isConnected = await _flutterThermalPrinterPlugin
+                          .connect(bleDevices[index]);
+                      log("Devices: $isConnected");
+                    },
+                    title: Text(bleDevices[index].name ?? 'No Name'),
+                    subtitle: Text(
+                        "VendorId: ${bleDevices[index].address} - Connected: ${bleDevices[index].isConnected}"),
+                    trailing: IconButton(
+                      icon: const Icon(Icons.connect_without_contact),
+                      onPressed: () async {
+                        final profile = await CapabilityProfile.load();
+                        final generator = Generator(PaperSize.mm80, profile);
+                        List<int> bytes = [];
+                        bytes += generator.text('Hello World');
+                        bytes +=
+                            generator.text("|||| FLUTTER THERMAL PRINTER ||||");
+                        bytes += generator.feed(2);
+                        bytes += generator.cut();
+                        await _flutterThermalPrinterPlugin.printData(
+                          bleDevices[index],
+                          bytes,
+                          longData: true,
                         );
                       },
                     ),
-                  ),
-                  Expanded(
-                    child: ListView.builder(
-                      itemCount: bleDevices.length,
-                      itemBuilder: (context, index) {
-                        return ListTile(
-                          onTap: () async {
-                            final isConnected =
-                                await _flutterThermalPrinterPlugin
-                                    .connect(bleDevices[index]);
-                            log("Devices: $isConnected");
-                          },
-                          title: Text(bleDevices[index].platformName),
-                          subtitle: Text(
-                              "VendorId: ${bleDevices[index].remoteId.str} - Connected: ${bleDevices[index].isConnected}"),
-                          trailing: IconButton(
-                            icon: const Icon(Icons.connect_without_contact),
-                            onPressed: () async {
-                              final profile = await CapabilityProfile.load();
-                              final generator =
-                                  Generator(PaperSize.mm80, profile);
-                              List<int> bytes = [];
-                              bytes += generator.text('Hello World');
-                              bytes += generator
-                                  .text("|||| FLUTTER THERMAL PRINTER ||||");
-                              bytes += generator.feed(2);
-                              bytes += generator.cut();
-                              await _flutterThermalPrinterPlugin.printData(
-                                bleDevices[index],
-                                bytes,
-                              );
-                            },
-                          ),
-                        );
-                      },
-                    ),
-                  ),
-                ],
+                  );
+                },
               ),
             ),
           ],